backend/requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv